Extract ApiExpectionMappers to helper class
