+++ basyx.aasregistry/basyx.aasregistry-feature-hierarchy/src/main/java/org/eclipse/digitaltwin/basyx/aasregistry/feature/hierarchy/mappers/AasRegistryModelMapper.java
@@ -23,15 +23,21 @@
  * SPDX-License-Identifier: MIT
  ******************************************************************************/
 
-package org.eclipse.digitaltwin.basyx.aasregistry.feature.hierarchy;
+package org.eclipse.digitaltwin.basyx.aasregistry.feature.hierarchy.mappers;
 
+import org.eclipse.digitaltwin.basyx.aasregistry.client.ApiException;
+import org.eclipse.digitaltwin.basyx.aasregistry.feature.hierarchy.HierarchalAasRegistryStorage;
 import org.eclipse.digitaltwin.basyx.aasregistry.model.AssetAdministrationShellDescriptor;
 import org.eclipse.digitaltwin.basyx.aasregistry.model.SubmodelDescriptor;
+import org.eclipse.digitaltwin.basyx.aasregistry.service.errors.AasDescriptorNotFoundException;
+import org.eclipse.digitaltwin.basyx.aasregistry.service.errors.SubmodelNotFoundException;
+import org.springframework.http.HttpStatus;
+import org.springframework.http.HttpStatusCode;
 
 import com.fasterxml.jackson.databind.ObjectMapper;
 
 /**
- * AasRegistryModelMapper
+ * Maps objects conversions for {@link HierarchalAasRegistryStorage}
  *
  * @author mateusmolina
  *
@@ -40,35 +46,42 @@ public final class AasRegistryModelMapper {
 
 	private static final ObjectMapper objectMapper = new ObjectMapper();
 
+	private AasRegistryModelMapper() {
+	}
+
 	public static org.eclipse.digitaltwin.basyx.aasregistry.client.model.AssetAdministrationShellDescriptor mapEqModel(AssetAdministrationShellDescriptor aasRegistryDescriptor) {
-		try {
 			return objectMapper.convertValue(aasRegistryDescriptor, org.eclipse.digitaltwin.basyx.aasregistry.client.model.AssetAdministrationShellDescriptor.class);
-		} catch (Exception e) {
-			throw new RuntimeException("Conversion error", e);
-		}
 	}
 
 	public static AssetAdministrationShellDescriptor mapEqModel(org.eclipse.digitaltwin.basyx.aasregistry.client.model.AssetAdministrationShellDescriptor aasRegistryDescriptor) {
-		try {
 			return objectMapper.convertValue(aasRegistryDescriptor, AssetAdministrationShellDescriptor.class);
-		} catch (Exception e) {
-			throw new RuntimeException("Conversion error", e);
-		}
 	}
 
 	public static org.eclipse.digitaltwin.basyx.aasregistry.client.model.SubmodelDescriptor mapEqModel(SubmodelDescriptor smRegistryDescriptor) {
-		try {
 			return objectMapper.convertValue(smRegistryDescriptor, org.eclipse.digitaltwin.basyx.aasregistry.client.model.SubmodelDescriptor.class);
-		} catch (Exception e) {
-			throw new RuntimeException("Conversion error", e);
-		}
 	}
 
 	public static SubmodelDescriptor mapEqModel(org.eclipse.digitaltwin.basyx.aasregistry.client.model.SubmodelDescriptor smRegistryDescriptor) {
-		try {
 			return objectMapper.convertValue(smRegistryDescriptor, SubmodelDescriptor.class);
-		} catch (Exception e) {
-			throw new RuntimeException("Conversion error", e);
-		}
 	}
+
+	public static RuntimeException mapApiException(ApiException e, String aasDescriptorId) {
+		if (HttpStatusCode.valueOf(e.getCode()).equals(HttpStatus.NOT_FOUND))
+			return new AasDescriptorNotFoundException(aasDescriptorId);
+
+		return new RuntimeException(e);
+	}
+
+	public static RuntimeException mapApiException(ApiException e, String aasDescriptorId, String smId) {
+		if (HttpStatusCode.valueOf(e.getCode()).equals(HttpStatus.NOT_FOUND) && checkIfSubmodelNotFound(e, aasDescriptorId, smId))
+			return new SubmodelNotFoundException(aasDescriptorId, smId);
+
+		return mapApiException(e, aasDescriptorId);
+	}
+
+	private static boolean checkIfSubmodelNotFound(ApiException e, String aasDescriptorId, String smId) {
+		SubmodelNotFoundException expectedException = new SubmodelNotFoundException(aasDescriptorId, smId);
+		return e.getMessage().contains(expectedException.getReason());
+	}
+
 }
